requirements.txt
numpy
pandas
scikit-learn
psutil
pytest
aiohttp
prometheus_client
opentelemetry-sdk
opentelemetry-exporter-otlp
protobuf

# Optional extras
xgboost; extra == "xgboost"
optuna; extra == "optuna"
stable-baselines3; extra == "rl"
